+++ core/src/mast/serialization/tests.rs
@@ -109,6 +109,7 @@ fn confirm_operation_and_decorator_structure() {
         Operation::HornerExt => (),
         Operation::EvalCircuit => (),
         Operation::Emit => (),
+        Operation::LogPrecompile => (),
     };
 
     match Decorator::Trace(0) {