+++ stdlib/tests/crypto/stark/mod.rs
@@ -90,7 +90,12 @@ pub fn generate_recursive_verifier_data(
     let program_info = ProgramInfo::from(program);
 
     // build public inputs and generate the advice data needed for recursive proof verification
-    let pub_inputs = PublicInputs::new(program_info, stack_inputs, stack_outputs);
+    let pub_inputs = PublicInputs::new(
+        program_info,
+        stack_inputs,
+        stack_outputs,
+        miden_core::precompile::PrecompileSponge::new(),
+    );
     let (_, proof, _precompile_requests) = proof.into_parts();
     Ok(generate_advice_inputs(proof, pub_inputs).unwrap())
 }